pyproject.toml
[tool.poetry]
name = "acapy-plugin-pickup"
version = "0.1.0"
description = "Pickup Protocol for ACA-Py"
authors = ["Daniel Bluhm <dbluhm@pm.me>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.6.9,<4.0"
aries-cloudagent = {git="https://github.com/dbluhm/aries-cloudagent-python", rev="feature/transport-manager-access"}
#aries-cloudagent = "^0.6.0"
pydantic = "^1.8.1"
typing-extensions = "^3.7.4"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.12.0"
black = "^20.8b1"
flake8 = "^3.9.1"
pytest = "^6.2.3"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
